Clean up documentation crate implementation
