+++ compiler-lsp/documentation/Cargo.toml
@@ -9,9 +9,7 @@ checking = { version = "0.1.0", path = "../../compiler-core/checking" }
 documenting = { version = "0.1.0", path = "../../compiler-core/documenting" }
 files = { version = "0.1.0", path = "../../compiler-core/files" }
 indexing = { version = "0.1.0", path = "../../compiler-core/indexing" }
-itertools = "0.14.0"
 lowering = { version = "0.1.0", path = "../../compiler-core/lowering" }
-rayon = "1.12.0"
 serde = { version = "1.0.228", features = ["derive"] }
 thiserror = "2.0.17"
 ts-rs = { version = "12.0.1", default-features = false, features = ["serde-compat"] }
